requirements.txt
# Camera
depthai>=2.24.0,<3.0.0
opencv-python>=4.8.0
numpy>=1.24.0
blobconverter>=1.2.0

# AI - Multi-provider support
anthropic>=0.18.0          # Claude (Anthropic)
openai>=1.12.0             # OpenAI (ChatGPT) and OpenRouter
google-generativeai>=0.3.0 # Google Gemini
python-dotenv>=1.0.0
Pillow>=10.0.0

# Speech
pyttsx3>=2.90
SpeechRecognition>=3.10.0
PyAudio>=0.2.13
